s3: libsmb: Change cli_smb2_setatr() to use 32-bit attributes.
The SMB_SET_FILE_BASIC_INFO info level this uses in SMB2
sets 32-bit attributes, so don't use SSVAL, use SIVAL.

No change needed in callers as implicit casts work fine.

+++ source3/libsmb/cli_smb2_fnum.h
@@ -197,7 +197,7 @@ NTSTATUS cli_smb2_setpathinfo(struct cli_state *cli,
 			const DATA_BLOB *p_in_data);
 NTSTATUS cli_smb2_setatr(struct cli_state *cli,
 			const char *fname,
-			uint16_t attr,
+			uint32_t attr,
 			time_t mtime);
 NTSTATUS cli_smb2_setattrE(struct cli_state *cli,
                         uint16_t fnum,
